+++ src/main/kotlin/ecommerce/auth/controller/AuthController.kt
@@ -12,8 +12,6 @@ import jakarta.validation.Valid
 import jakarta.validation.ValidationException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.validation.FieldError
-import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
@@ -54,17 +52,6 @@ class AuthController(
     @ExceptionHandler(ValidationException::class)
     fun handleValidationException(ex: ValidationException): ResponseEntity<Map<String, String>> {
         return ResponseEntity(mapOf("error" to (ex.message ?: "Validation error")), HttpStatus.FORBIDDEN)
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException::class)
-    fun handleValidationExceptions(ex: MethodArgumentNotValidException): ResponseEntity<Map<String, String>> {
-        val errors = mutableMapOf<String, String>()
-        ex.bindingResult.allErrors.forEach { error ->
-            val fieldName = (error as FieldError).field
-            val errorMessage = error.defaultMessage ?: "Invalid value"
-            errors[fieldName] = errorMessage
-        }
-        return ResponseEntity(errors, HttpStatus.BAD_REQUEST)
     }
 
     @ExceptionHandler(AuthorizationException::class)
