+++ src/main/kotlin/ecommerce/product/controller/ProductApiController.kt
@@ -6,8 +6,6 @@ import jakarta.validation.Valid
 import jakarta.validation.ValidationException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.validation.FieldError
-import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.GetMapping
@@ -49,17 +47,6 @@ class ProductApiController(private val productRepository: ProductRepository) {
     ): ResponseEntity<Void> {
         productRepository.delete(id)
         return ResponseEntity.noContent().build()
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException::class)
-    fun handleValidationExceptions(ex: MethodArgumentNotValidException): ResponseEntity<Map<String, String>> {
-        val errors = mutableMapOf<String, String>()
-        ex.bindingResult.allErrors.forEach { error ->
-            val fieldName = (error as FieldError).field
-            val errorMessage = error.defaultMessage ?: "Invalid value"
-            errors[fieldName] = errorMessage
-        }
-        return ResponseEntity(errors, HttpStatus.BAD_REQUEST)
     }
 
     @ExceptionHandler(ValidationException::class)
